split functions into core and helpers
