+++ esri_heightmap/__init__.py
@@ -1,3 +1,3 @@
 # initialization script for importing as package
 print('initializing...')
-from .functions import *
+from .core import *
