+++ esri_heightmap/helpers.py
@@ -1,40 +1,13 @@
+# more functions from the package
+
 import numpy as np
 from PIL import Image
-import re
-
-## gets the metadata from a *.asc file and returns it as a friendly dictionary object
-def getAscMeta(fileObj):
-    # borrowed from https://github.com/domlysz/BlenderGIS, in turn derived from https://github.com/hrbaer/Blender-ASCII-Grid-Import - no license included in either; replace with bespoke code in near future to avoid conflict
-    meta_re = re.compile(r'^([^\s]+)\s+([^\s]+)$') # strange regex magic
-    meta = {}
-    for i in range(6):
-        line = fileObj.readline()
-        m = meta_re.match(line)
-        if m:
-            meta[m.group(1).lower()] = float(m.group(2))
-    return meta
 
 ## resizes a NumPy array by a given scale factor
 def arrayResize(array,scale):
         newSizeX = int(array.shape[1]*scale)
         newSizeY = int(array.shape[0]*scale)
         return np.array(Image.fromarray(array).resize((newSizeX,newSizeY),resample=Image.BILINEAR))
-
-
-## superimposes an asc array onto a primary array using metadata
-def overlayCell(mainArray, originOffset, newArray, ascMeta, finalResolution):            
-    newArray[newArray==ascMeta['nodata_value']] = np.nan # eliminate nodata_value into nice NaNs
-    
-    if ascMeta['cellsize'] != finalResolution: # if cell size is different, we must resize
-        newArray = arrayResize(newArray,ascMeta['cellsize']/finalResolution)
-        
-    # get offset coordinates - where to superimpose the array
-    newArrayOffset = (int(mainArray.shape[0]-(ascMeta['yllcorner']-originOffset[1])*(1/finalResolution)-newArray.shape[1]),int((ascMeta['xllcorner']-originOffset[0])*(1/finalResolution))) # very complicated code to find position of new array
-    
-    # perform the superimposition (*fancyword*)
-    mainArray[newArrayOffset[0]:newArrayOffset[0]+int(newArray.shape[1]),newArrayOffset[1]:newArrayOffset[1]+int(newArray.shape[0])] = newArray  # this should be neater
-    return mainArray
-
 
 ## eliminates NaNs (don't take too personally) by attempting to find data in secondary array (data of lower resolution) and other methods (_future_)
 # there may be a NumPy function to do this already (I couldn't find one) - if there is, please use it instead OR update the source and create a pull request OR tell me!
@@ -50,7 +23,6 @@ def NaNReplace(primaryArray,secondaryArray):
                 # at this point, an "else" clause should be added to peform some kind of interpolaion patchwork for remaining NaNs
     return primaryArray
 
-
 ## use given number of bits for array, losing distrubution information
 def normalizeArray(array,bitDepth):
     # normalize the heights into _bitDepth_ integers
@@ -61,7 +33,7 @@ def normalizeArray(array,bitDepth):
 ## tifffile wrapper
 def exportTiff(normalizedArray,file):
     try:
-        import tifffile as tiff
+        import tifffile as tiff # optional dependency
         tiff.imwrite(file,normalizedArray)
     except ImportError:
         print('tifffile required for tiff export')
